Re-organize and format code
