--- bbx_audio/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "bbx_audio"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-rodio = "0.20.1"
