--- bbx_draw/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "bbx_draw"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-bbx_dsp = { path = "../bbx_dsp" }
-
-nannou = "0.19.0"
