--- bbx_draw/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod chart;
-pub mod context;
-pub mod utilities;
-
-pub use chart::{draw_chart, ChartConfiguration};
-pub use context::DisplayContext;
-pub use utilities::{map_sample_data_to_point2, scale_number};
