+++ bbx_dsp/Cargo.toml
@@ -4,9 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-bbx_buffer = { path = "../bbx_buffer" }
-bbx_file = { path = "../bbx_file" }
-bbx_sample = { path = "../bbx_sample" }
-
-rand = "0.8.5"
-thiserror = "1.0.61"
+rodio = "0.20.1"
