--- bbx_dsp/src/constants.rs
@@ -1,9 +1,0 @@
-use crate::context::Context;
-
-/// The default context in which to evaluate a `Graph`.
-pub const DEFAULT_CONTEXT: Context = Context {
-    sample_rate: 44100,
-    num_channels: 2,
-    buffer_size: 128,
-    max_num_graph_nodes: 1024,
-};
