--- bbx_dsp/src/effectors/mod.rs
@@ -1,5 +1,0 @@
-pub mod amplifier;
-pub mod filter;
-pub mod flanger;
-pub mod mixer;
-pub mod overdrive;
