+++ bbx_dsp/src/lib.rs
@@ -1,16 +1,7 @@
-#![feature(test)]
-extern crate test;
-
-pub mod effectors;
-pub mod generators;
-
-pub mod constants;
+pub mod block;
+pub mod blocks;
 pub mod context;
-pub mod effector;
-pub mod error;
-pub mod generator;
 pub mod graph;
-pub mod node;
-pub mod operation;
-pub mod process;
-pub mod utils;
+pub mod parameter;
+pub mod sample;
+pub mod waveform;
