+++ bbx_dsp/src/sample.rs
@@ -1,0 +1,51 @@
+use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign};
+
+pub trait Sample:
+    Copy
+    + Clone
+    + Send
+    + Sync
+    + 'static
+    + Add<Output = Self>
+    + Sub<Output = Self>
+    + Mul<Output = Self>
+    + Div<Output = Self>
+    + AddAssign
+    + SubAssign
+    + MulAssign
+    + DivAssign
+    + PartialOrd
+    + PartialEq
+{
+    const ZERO: Self;
+    const ONE: Self;
+
+    fn from_f64(value: f64) -> Self;
+    fn to_f64(self) -> f64;
+}
+
+impl Sample for f32 {
+    const ZERO: Self = 0.0;
+    const ONE: Self = 1.0;
+
+    fn from_f64(value: f64) -> Self {
+        value as f32
+    }
+
+    fn to_f64(self) -> f64 {
+        self as f64
+    }
+}
+
+impl Sample for f64 {
+    const ZERO: Self = 0.0;
+    const ONE: Self = 1.0;
+
+    fn from_f64(value: f64) -> Self {
+        value
+    }
+
+    fn to_f64(self) -> f64 {
+        self
+    }
+}
