+++ bbx_dsp/src/waveform.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone)]
+pub enum Waveform {
+    Sine,
+}
