+++ bbx_midi/Cargo.toml
@@ -5,4 +5,4 @@ edition = "2021"
 
 [dependencies]
 midir = "0.10.0"
-thiserror = "1.0.63"
+thiserror = "2.0.6"
