--- bbx_sample/Cargo.toml
@@ -1,4 +1,0 @@
-[package]
-name = "bbx_sample"
-version = "0.1.0"
-edition = "2021"
