--- bbx_sample/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod float;
-pub mod sample;
