+++ bbx_sandbox/Cargo.toml
@@ -4,9 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-bbx_buffer = { path = "../bbx_buffer" }
 bbx_dsp = { path = "../bbx_dsp" }
 bbx_midi = { path = "../bbx_midi" }
 
 rand = "0.8.5"
-rodio = "0.18.1"
+rodio = "0.20.1"
