--- bbx_sandbox/src/constants.rs
@@ -1,1 +1,0 @@
-pub const PLAYTIME_DURATION_SECONDS: usize = 3;
