+++ bbx_sandbox/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod constants;
 pub mod player;
 pub mod signal;
